STR2726 fix. proper cast should be int as specified in isspace documentation, not unsigned char
git-svn-id: file:///fltk/svn/fltk/branches/branch-1.3@9363 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ fluid/Fl_Type.cxx
@@ -594,9 +594,9 @@ int storestring(const char *n, const char * & p, int nostrip) {
   undo_checkpoint();
   int length = 0;
   if (n) { // see if blank, strip leading & trailing blanks
-    if (!nostrip) while (isspace(*n)) n++;
+    if (!nostrip) while (isspace((int) *n)) n++;
     const char *e = n + strlen(n);
-    if (!nostrip) while (e > n && isspace(*(e-1))) e--;
+    if (!nostrip) while (e > n && isspace((int)*(e-1))) e--;
     length = e-n;
     if (!length) n = 0;
   }
